requirements.txt
Package                  Version
------------------------ ---------
aiohttp                  3.9.5
aiosignal                1.3.1
annotated-types          0.7.0
anyio                    4.4.0
attrs                    23.2.0
blinker                  1.8.2
certifi                  2024.7.4
charset-normalizer       3.3.2
click                    8.1.7
filelock                 3.15.4
Flask                    3.0.3
frozenlist               1.4.1
fsspec                   2024.6.1
greenlet                 3.0.3
h11                      0.14.0
httpcore                 1.0.5
httpx                    0.27.0
huggingface-hub          0.24.2
idna                     3.7
itsdangerous             2.2.0
Jinja2                   3.1.4
jsonpatch                1.33
jsonpointer              3.0.0
langchain                0.2.11
langchain-core           0.2.24
langchain-ollama         0.1.0
langchain-text-splitters 0.2.2
langsmith                0.1.93
MarkupSafe               2.1.5
multidict                6.0.5
numpy                    1.26.4
ollama                   0.3.0
orjson                   3.10.6
packaging                24.1
pip                      24.2
pydantic                 2.8.2
pydantic_core            2.20.1
pyodbc                   5.1.0
python-dotenv            1.0.1
PyYAML                   6.0.1
regex                    2024.7.24
requests                 2.32.3
safetensors              0.4.3
setuptools               72.1.0
sniffio                  1.3.1
SQLAlchemy               2.0.31
tenacity                 8.5.0
tokenizers               0.19.1
tqdm                     4.66.4
transformers             4.43.3
typing_extensions        4.12.2
urllib3                  2.2.2
uuid                     1.30
Werkzeug                 3.0.3
wheel                    0.43.0
yarl                     1.9.4
